requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.14
hypercorn==0.16.0
greenlet==3.2.3
h11==0.16.0
httptools==0.6.4
idna==3.10
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1

# Testing
pytest==8.3.4
pytest-cov==6.0.0
httpx==0.28.1
